Split snapshot tests into test_migration_snapshots.py
Separate concerns:

  - test_versioned_migrations.py keeps framework plumbing tests —
    registry validation, fresh init, version gating, legacy
    bootstrap, numbered-migration runner, SqlFileMigration wiring,
    concurrency, PRAGMA heuristic.
  - test_migration_snapshots.py is the new home for the pair-based
    snapshot harness, the ``test_latest_snapshot_schema_matches_schema_sql``
    sanity check, and the ``_regenerate_snapshots`` one-shot
    generator (including the seed functions).

This is the right split: the snapshot tests exist to verify
migrations, while the framework tests verify the framework. When
writing a new migration you'll touch test_migration_snapshots.py
(add a set, regenerate baselines) but not test_versioned_migrations.py.

No behavioural changes; everything still runs under
``uv run --group dev pytest``.
