+++ compute_space/tests/test_versioned_migrations.py
@@ -1,36 +1,30 @@
-"""Tests for the versioned migration framework.
+"""Tests for the versioned migration framework itself.
+
+Scope: framework plumbing — base class, registry validation, runner
+behaviour, legacy bootstrap, fresh init, concurrency, the PRAGMA
+heuristic. The migrations themselves are covered by the snapshot
+tests in ``test_migration_snapshots.py``.
 
 Covers:
   - schema_version metadata table (REQ-VER-1..5)
   - legacy v0 -> v1 bootstrap via the existing migrate() (REQ-LEG-*)
   - Migration base class + SqlFileMigration (REQ-MF-*)
   - Registry validation: contiguous, starts at 2 (REQ-REG-*)
-  - Runner behavior: lock, apply, atomic version bump, rollback (REQ-RUN-*)
+  - Runner behaviour: lock, apply, atomic version bump, rollback (REQ-RUN-*)
   - Fresh-DB init via schema.sql, stamps to highest (REQ-INIT-*)
-  - Pair-based snapshot tests over tests/snapshots/<set>/v*.sql
-    (REQ-TEST-1, REQ-TEST-2, REQ-TEST-3, REQ-TEST-4)
   - Legacy bootstrap fixture test (REQ-TEST-5)
   - Concurrency: two concurrent startups (REQ-TEST-6)
   - PRAGMA foreign_keys heuristic warning (REQ-TEST-7)
-
-Snapshot format: each ``snapshots/<set>/vNNNN.sql`` is the output of
-``sqlite3.Connection.iterdump()``, with ``sqlite_sequence`` lines
-filtered out (environment-dependent autoincrement state). On mismatch
-the harness writes the actual dump to ``<name>.sql.new`` next to the
-expected file and fails with a unified diff. Rename the ``.new`` to
-``.sql`` to accept.
 """
 
 from __future__ import annotations
 
-import difflib
 import importlib
 import inspect
 import re
 import sqlite3
 import subprocess
 import sys
-import tempfile
 import time
 import warnings as _warnings
 from pathlib import Path
@@ -632,271 +626,6 @@ class TestConcurrency:
             assert row is not None
         finally:
             db.close()
-
-
-# --------------------------------------------------------------------------- #
-# Snapshot harness (REQ-TEST-1..4)
-# --------------------------------------------------------------------------- #
-
-
-SNAPSHOTS_DIR = Path(__file__).resolve().parent / "snapshots"
-_VERSION_FILE_RE = re.compile(r"^v(\d+)\.sql$")
-
-
-def _snapshot_sets() -> list[tuple[str, list[Path]]]:
-    """Discover ``snapshots/<set>/v*.sql`` groupings.
-
-    Returns a list of ``(set_name, sorted_files)`` for every subfolder
-    with at least one ``vNNNN.sql`` file.
-    """
-    if not SNAPSHOTS_DIR.exists():
-        return []
-    sets: list[tuple[str, list[Path]]] = []
-    for set_dir in sorted(SNAPSHOTS_DIR.iterdir()):
-        if not set_dir.is_dir():
-            continue
-        versioned: list[tuple[int, Path]] = []
-        for f in set_dir.iterdir():
-            m = _VERSION_FILE_RE.match(f.name)
-            if m:
-                versioned.append((int(m.group(1)), f))
-        versioned.sort()
-        if versioned:
-            sets.append((set_dir.name, [p for _, p in versioned]))
-    return sets
-
-
-def _version_of(path: Path) -> int:
-    m = _VERSION_FILE_RE.match(path.name)
-    assert m, f"not a vNNNN.sql snapshot: {path.name}"
-    return int(m.group(1))
-
-
-def _snapshot_pairs() -> list[tuple[str, Path, Path]]:
-    """Pairs of consecutive vN files *present* in each set (not consecutive
-    integers — if a set has v1 and v15 only, there is one pair (1, 15)).
-    """
-    out: list[tuple[str, Path, Path]] = []
-    for set_name, files in _snapshot_sets():
-        for a, b in zip(files, files[1:], strict=False):
-            out.append((set_name, a, b))
-    return out
-
-
-def _dump_for_snapshot(db_path: str) -> str:
-    """Render a DB to its canonical .sql snapshot form.
-
-    Uses :meth:`sqlite3.Connection.iterdump`, joined with newlines. The
-    only whitelisted exclusion is ``sqlite_sequence`` — its state is a
-    function of prior AUTOINCREMENT allocations and is environment-noise
-    for schema/data correctness checks.
-    """
-    db = sqlite3.connect(db_path)
-    try:
-        lines = [line for line in db.iterdump() if "sqlite_sequence" not in line]
-    finally:
-        db.close()
-    return "\n".join(lines) + "\n"
-
-
-def _write_new_and_fail(expected_path: Path, actual: str, msg_prefix: str) -> None:
-    """Write ``<name>.sql.new`` next to the expected file and fail with a diff.
-
-    Developer workflow: inspect the ``.sql.new``, rename to ``.sql`` once
-    the change is intentional. No --update-snapshots flag.
-    """
-    new_path = expected_path.with_suffix(expected_path.suffix + ".new")
-    new_path.write_text(actual)
-    rel_expected = expected_path.relative_to(SNAPSHOTS_DIR.parent)
-    rel_new = new_path.relative_to(SNAPSHOTS_DIR.parent)
-    if expected_path.exists():
-        expected = expected_path.read_text()
-        diff = "".join(
-            difflib.unified_diff(
-                expected.splitlines(keepends=True),
-                actual.splitlines(keepends=True),
-                fromfile=str(rel_expected),
-                tofile=str(rel_new),
-            )
-        )
-        pytest.fail(f"{msg_prefix}\nWrote actual dump to {rel_new}. Diff:\n{diff}")
-    else:
-        pytest.fail(
-            f"{msg_prefix}\nMissing golden {rel_expected}. Wrote generated dump to {rel_new}; "
-            "review and rename to accept."
-        )
-
-
-@pytest.mark.parametrize(
-    ("set_name", "va", "vb"),
-    _snapshot_pairs(),
-    ids=[f"{name}-v{_version_of(a)}-to-v{_version_of(b)}" for name, a, b in _snapshot_pairs()],
-)
-def test_snapshot_pair(set_name, va, vb, tmp_path):
-    """REQ-TEST-1/2/3: from ``<set>/vA.sql``, apply REGISTRY, must dump to ``<set>/vB.sql``."""
-    a_version = _version_of(va)
-    b_version = _version_of(vb)
-
-    db_path = str(tmp_path / f"{set_name}_v{a_version}_to_v{b_version}.db")
-    # Load the starting snapshot verbatim.
-    loader = sqlite3.connect(db_path)
-    try:
-        loader.executescript(va.read_text())
-    finally:
-        loader.close()
-
-    probe = sqlite3.connect(db_path)
-    try:
-        assert read_version(probe) == a_version, (
-            f"{va.relative_to(SNAPSHOTS_DIR.parent)} claims v{a_version} but loads as v{read_version(probe)}"
-        )
-    finally:
-        probe.close()
-
-    apply_migrations(db_path, registry=REGISTRY)
-
-    probe = sqlite3.connect(db_path)
-    try:
-        final_version = read_version(probe)
-    finally:
-        probe.close()
-    assert final_version == b_version, (
-        f"After applying REGISTRY starting at v{a_version}, expected v{b_version} but DB is at v{final_version}"
-    )
-
-    actual = _dump_for_snapshot(db_path)
-    expected = vb.read_text()
-    if actual != expected:
-        _write_new_and_fail(
-            vb,
-            actual,
-            f"Snapshot mismatch for set={set_name} pair=v{a_version}->v{b_version}",
-        )
-
-
-# --------------------------------------------------------------------------- #
-# Snapshot vs schema.sql sanity (REQ-TEST-4)
-# --------------------------------------------------------------------------- #
-
-
-@pytest.mark.parametrize(
-    ("set_name", "files"),
-    _snapshot_sets(),
-    ids=[name for name, _ in _snapshot_sets()],
-)
-def test_latest_snapshot_schema_matches_schema_sql(set_name, files, tmp_path):
-    """The latest snapshot in each set must have a schema equivalent to a
-    fresh ``schema.sql`` initialization.
-
-    Only the schema is compared — snapshots carry seed data whereas a
-    fresh init is empty; ``assert_schemas_equal`` ignores row contents.
-    """
-    latest = files[-1]
-    snap_path = str(tmp_path / f"{set_name}_latest.db")
-    loader = sqlite3.connect(snap_path)
-    try:
-        loader.executescript(latest.read_text())
-    finally:
-        loader.close()
-
-    fresh_path = str(tmp_path / f"{set_name}_fresh.db")
-    apply_migrations(fresh_path, registry=REGISTRY)
-
-    snap_db = sqlite3.connect(snap_path)
-    fresh_db = sqlite3.connect(fresh_path)
-    try:
-        assert_schemas_equal(get_schema_snapshot(fresh_db), get_schema_snapshot(snap_db))
-    finally:
-        snap_db.close()
-        fresh_db.close()
-
-
-# --------------------------------------------------------------------------- #
-# Snapshot generation
-# --------------------------------------------------------------------------- #
-
-
-def _seed_empty(_db: sqlite3.Connection) -> None:  # pragma: no cover - generator
-    pass
-
-
-def _seed_service_with_ports(db: sqlite3.Connection) -> None:  # pragma: no cover - generator
-    """Minimal realistic seed: one app, one port mapping, one service provider."""
-    db.execute(
-        "INSERT INTO apps (name, manifest_name, version, runtime_type, repo_path, "
-        "local_port, description, memory_mb, cpu_millicores, gpu, public_paths, "
-        "created_at, updated_at) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?)",
-        (
-            "orders",
-            "orders",
-            "1.0.0",
-            "serverfull",
-            "/repo/orders",
-            19100,
-            "Order service",
-            256,
-            500,
-            0,
-            "[]",
-            "2024-01-01T00:00:00",
-            "2024-01-01T00:00:00",
-        ),
-    )
-    db.execute(
-        "INSERT INTO app_port_mappings (app_name, label, container_port, host_port) VALUES (?,?,?,?)",
-        ("orders", "grpc", 8000, 19500),
-    )
-    db.execute(
-        "INSERT INTO service_providers (service_name, app_name) VALUES (?,?)",
-        ("payments", "orders"),
-    )
-
-
-_SEEDS = {
-    "empty": _seed_empty,
-    "service_with_ports": _seed_service_with_ports,
-}
-
-
-def _regenerate_snapshots(root: Path = SNAPSHOTS_DIR) -> None:  # pragma: no cover - generator
-    """Rebuild every ``<set>/vNNNN.sql`` from the live code + seed functions.
-
-    Not a test. Invoke as::
-
-        python -c 'from compute_space.tests.test_versioned_migrations \\
-            import _regenerate_snapshots; _regenerate_snapshots()'
-
-    Writes the entire snapshot tree in place. Review diffs before committing.
-    The harness's normal .sql.new workflow is preferred for small updates;
-    this generator is for initial bootstrap and wholesale resets.
-    """
-    root.mkdir(parents=True, exist_ok=True)
-
-    for set_name, seed in _SEEDS.items():
-        set_dir = root / set_name
-        set_dir.mkdir(parents=True, exist_ok=True)
-
-        with tempfile.TemporaryDirectory() as td:
-            td_path = Path(td)
-            # v1: bootstrap to v1 + seed.
-            v1_db = str(td_path / f"{set_name}_v1.db")
-            apply_migrations(v1_db, registry=[])
-            with sqlite3.connect(v1_db, isolation_level=None) as conn:
-                seed(conn)
-            (set_dir / "v0001.sql").write_text(_dump_for_snapshot(v1_db))
-
-            # Higher versions: start from the v1 dump, apply REGISTRY up to
-            # each registered version. This mirrors the pair-test flow.
-            for migration in REGISTRY:
-                target = migration.version
-                next_db = str(td_path / f"{set_name}_v{target}.db")
-                loader = sqlite3.connect(next_db)
-                try:
-                    loader.executescript((set_dir / "v0001.sql").read_text())
-                finally:
-                    loader.close()
-                apply_migrations(next_db, registry=REGISTRY)
-                (set_dir / f"v{target:04d}.sql").write_text(_dump_for_snapshot(next_db))
 
 
 # --------------------------------------------------------------------------- #
